feat: added user accounts fe
