+++ src/components/account/ProfileTab.tsx
@@ -1,0 +1,11 @@
+import ProfileWallet from './ProfileWallet'
+
+const ProfileTab = () => {
+  return (
+    <div className="w-3/4 h-full">
+      <ProfileWallet />
+    </div>
+  )
+}
+
+export default ProfileTab
