+++ src/pages/embed.tsx
@@ -48,7 +48,7 @@ export default function Embed() {
   return (
     <>
       <div className="overflow-x-hidden bg-brand-gray">
-        <div className="w-screen px-6 pt-10 pb-40 text-center mb-40  md:mb-5 text-white bg-cover bg-top-mobile md:bg-top-desktop">
+        <div className="w-screen px-6 pt-10 pb-40 mb-40 text-center text-white bg-cover md:mb-5 bg-top-mobile md:bg-top-desktop">
           <div>
             <h2 className="text-3xl md:text-6xl font-gilroy-bold">
               Tuning into what <span className="text-brand-blue">matters</span>
@@ -57,22 +57,22 @@ export default function Embed() {
               Allow people to voice their trust by embeding ideamarket social
             </p>
           </div>
-          <div className="flex  mt-14 flex-col items-center justify-center  text-md md:text-3xl font-gilroy-bold md:flex-row">
+          <div className="flex flex-col items-center justify-center mt-14 text-md md:text-3xl font-gilroy-bold md:flex-row">
             <form
               className="w-full max-w-lg"
               action="POST"
               onSubmit={createEmbed}
             >
-              <div className="flex flex-wrap justify-center -mx-3 mb-6">
-                <div className="w-full md:w-1/3 px-3 mb-6 md:mb-0">
+              <div className="flex flex-wrap justify-center mb-6 -mx-3">
+                <div className="w-full px-3 mb-6 md:w-1/3 md:mb-0">
                   <label
-                    className="block uppercase tracking-wide text-gray-400 text-xs font-bold mb-2"
+                    className="block mb-2 text-xs font-bold tracking-wide text-gray-400 uppercase"
                     htmlFor="grid-first-name"
                   >
                     Name
                   </label>
                   <input
-                    className="appearance-none block w-full text-lg bg-gray-200 text-gray-700 border rounded py-2 px-4 mb-2 leading-tight focus:outline-none focus:bg-white"
+                    className="block w-full px-4 py-2 mb-2 text-lg leading-tight text-gray-700 bg-gray-200 border rounded appearance-none focus:outline-none focus:bg-white"
                     id="tagname"
                     name="tagname"
                     type="text"
@@ -81,9 +81,9 @@ export default function Embed() {
                     onChange={({ target }) => setTagname(target.value)}
                   ></input>
                 </div>
-                <div className="w-full md:w-1/3 px-3">
+                <div className="w-full px-3 md:w-1/3">
                   <label
-                    className="block uppercase tracking-wide text-gray-400 text-xs font-bold mb-2"
+                    className="block mb-2 text-xs font-bold tracking-wide text-gray-400 uppercase"
                     htmlFor="grid-last-name"
                   >
                     Market
@@ -92,7 +92,7 @@ export default function Embed() {
                   <Select
                     options={selectMarketValues}
                     formatOptionLabel={selectMarketFormat}
-                    className="block appearance-none w-full bg-gray-200 border border-gray-200 text-gray-700 text-lg  rounded leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="block w-full text-lg leading-tight text-gray-700 bg-gray-200 border border-gray-200 rounded appearance-none focus:outline-none focus:bg-white focus:border-gray-500"
                     defaultInputValue="Twitter"
                     onChange={(entry) => {
                       setMarket((entry as any).value)
@@ -101,7 +101,7 @@ export default function Embed() {
                 </div>
               </div>
               <button
-                className="py-2 -mt-2 ml-5 text-lg font-bold text-white rounded-md w-44 font-sf-compact-medium bg-brand-blue hover:bg-blue-800"
+                className="py-2 ml-5 -mt-2 text-lg font-bold text-white rounded-md w-44 font-sf-compact-medium bg-brand-blue hover:bg-blue-800"
                 type="submit"
               >
                 <div className="flex flex-row items-center justify-center">
@@ -111,7 +111,7 @@ export default function Embed() {
             </form>
           </div>
 
-          <div className="mx-5 m-20 grid place-content-center">
+          <div className="grid m-20 mx-5 place-content-center">
             <iframe
               src={`http://localhost:3000/iframe/${market}/${tagName}`}
               width={ewidth}
@@ -119,21 +119,21 @@ export default function Embed() {
               title="preview"
             ></iframe>
 
-            <div className="flex justify-center items-center sm:mt-24 mb:mt-0">
+            <div className="flex items-center justify-center sm:mt-24 mb:mt-0">
               <dialog
                 open
-                className="rounded-2xl overflow-hidden p-0 w-auto max-w-7xl md:mx-auto md:w-1/2 shadow-lg m-8 mt-52"
+                className="w-auto p-0 m-8 overflow-hidden shadow-lg rounded-2xl max-w-7xl md:mx-auto md:w-1/2 mt-52"
               >
                 <div className="flex flex-col lg:flex-row">
                   <div className="w-full">
                     <div className="p-6">
-                      <div className="flex justify-between items-start">
-                        <h3 className="text-xl font-bold mb-6">
+                      <div className="flex items-start justify-between">
+                        <h3 className="mb-6 text-xl font-bold">
                           Generated code
                         </h3>
 
                         <button
-                          className="outline-none focus:outline-none border-gray-200 w-8 h-8 hover:text-green-500 active:bg-gray-50"
+                          className="w-8 h-8 border-gray-200 outline-none focus:outline-none hover:text-green-500 active:bg-gray-50"
                           onClick={() => {
                             setCopyDone(true)
                           }}
@@ -147,7 +147,7 @@ export default function Embed() {
                       </div>
                       <div className="overflow-hidden rounded-md">
                         <textarea
-                          className="w-full px-3 py-1 border border-gray-200  rounded-md focus:outline-none resize-none"
+                          className="w-full px-3 py-1 border border-gray-200 rounded-md resize-none focus:outline-none"
                           value={embed}
                           rows={2}
                           onChange={(e) => {
@@ -166,6 +166,7 @@ export default function Embed() {
     </>
   )
 }
+
 Embed.layoutProps = {
   Layout: DefaultLayout,
 }
