+++ tailwind.config.js
@@ -86,6 +86,8 @@ module.exports = {
       backgroundImage: (theme) => ({
         'top-mobile': "url('/topbg-mobile.svg')",
         'top-desktop': "url('/topbg.svg')",
+        'top-desktop-new': "url('/topbg-new.png')",
+        'top-mobile-new': "url('/topbg-new.png')",
       }),
       boxShadow: {
         home: '0 1px 5px 0 rgba(0, 0, 0, 0.05)',
